+++ domain/src/main/java/com/example/domain/model/diary/NutritionData.kt
@@ -2,7 +2,7 @@ package com.example.domain.model.diary
 
 data class NutritionData(
     val calories: Int = 0,
-    val carb: Int = 0,
-    val protein: Int = 0,
-    val fat: Int = 0
+    val carb: Float = 0f,
+    val protein: Float = 0f,
+    val fat: Float = 0f
 )
